PR fixes. Remove walrus operator and use the `start_as_current_span`
