--- instrumentation/opentelemetry-instrumentation-boto3sqs/tests/test_getter.py
@@ -1,63 +1,0 @@
-# Copyright The OpenTelemetry Authors
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-# pylint: disable=no-name-in-module
-
-from unittest import TestCase
-
-from opentelemetry.instrumentation.boto3sqs import (
-    OPENTELEMETRY_ATTRIBUTE_IDENTIFIER,
-    Boto3SQSGetter,
-)
-
-
-class TestBoto3SQSGetter(TestCase):
-    def setUp(self) -> None:
-        self.getter = Boto3SQSGetter()
-
-    def test_get_none(self) -> None:
-        carrier = {}
-        value = self.getter.get(carrier, "test")
-        self.assertIsNone(value)
-
-    def test_get_value(self) -> None:
-        key = "test"
-        value = "value"
-        carrier = {
-            f"{OPENTELEMETRY_ATTRIBUTE_IDENTIFIER}{key}": {
-                "StringValue": value,
-                "DataType": "String",
-            }
-        }
-        val = self.getter.get(carrier, key)
-        self.assertEqual(val, [value])
-
-    def test_keys(self):
-        key1 = "test1"
-        value1 = "value1"
-        key2 = "test2"
-        value2 = "value2"
-        carrier = {
-            f"{OPENTELEMETRY_ATTRIBUTE_IDENTIFIER}{key1}": {
-                "StringValue": value1,
-                "DataType": "String",
-            },
-            key2: {"StringValue": value2, "DataType": "String"},
-        }
-        keys = self.getter.keys(carrier)
-        self.assertEqual(keys, [key1, key2])
-
-    def test_keys_empty(self):
-        keys = self.getter.keys({})
-        self.assertEqual(keys, [])
